Arreglando errores, rutas, favicons, etc..

+++ usuario/eliminar_provincia.php
@@ -1,7 +1,14 @@
 <?php
 include("../php/conexion.php");
-
+session_start();
 $id = $_GET['id'] ?? null;
+$usuario = $_SESSION['id_rol'];
+$conexion->query("SET @usuario_app = '$usuario'");
+
+if (!$id) {
+    header("Location: provincias.php");
+    exit;
+}
 
 if ($id) {
     $stmt = $conexion->prepare("DELETE FROM provincia WHERE id_provincia = ?");
